Fixed compiler warnings, mainly mismatched string formats. Updated ObjectiveFlickr submodule to latest commit on master. Updated project settings to use LLDB.

+++ IMBObjectViewController.m
@@ -1875,7 +1875,7 @@ NSString* const IMBObjectViewControllerSegmentedControlKey = @"SegmentedControl"
 							}
 							
 							[thisItem setDataProvider:self forTypes:whichTypes];
-							[thisItem setString:[NSString stringWithFormat:@"%d", thisIndex] forType:kIMBPrivateItemIndexPasteboardType];
+							[thisItem setString:[NSString stringWithFormat:@"%lu", (unsigned long)thisIndex] forType:kIMBPrivateItemIndexPasteboardType];
 							[thisItem setString:thisObject.name forType:kIMBPublicTitleListPasteboardType];
 							[thisItem setPropertyList:thisObject.metadata forType:kIMBPublicMetadataListPasteboardType];
 							[thisItem setData:promiseData forType:kIMBPasteboardTypeObjectsPromise];
